+++ src/appl/rmcdhf90_mem_mpi/systemmem.f90
@@ -21,7 +21,7 @@
 !MemTotal:       264566412 kB
 !MemTotal:       1056570236 kB
       REAL(DOUBLE) :: SYSTEMMem
-      INTEGER      :: SUMMem, LENG
+      INTEGER*8      :: SUMMem, LENG
       CHARACTER(LEN=128) :: LINE,CMem
 !-----------------------------------------------
 !
